Tolti i cast che tanto non servivano

+++ pandos/phase3/sysSupport.c
@@ -58,12 +58,12 @@ void trapExcHandler(support_t *currSupStructPTR)
             swap_pool_table[i].sw_asid = -1;
     }
     // Sblocco il traffico
-    SYSCALL(VERHOGEN, (int)&printer_sem[asid], 0, 0);
-    SYSCALL(VERHOGEN, (int)&flash_sem[asid], 0, 0);
-    SYSCALL(VERHOGEN, (int)&term_w_sem[asid], 0, 0);
-    SYSCALL(VERHOGEN, (int)&term_r_sem[asid], 0, 0);
-    SYSCALL(VERHOGEN, (int)&swapSemaphore, 0, 0);
-    SYSCALL(VERHOGEN, (int)&mainSemaphore, 0, 0);
+    SYSCALL(VERHOGEN, &printer_sem[asid], 0, 0);
+    SYSCALL(VERHOGEN, &flash_sem[asid], 0, 0);
+    SYSCALL(VERHOGEN, &term_w_sem[asid], 0, 0);
+    SYSCALL(VERHOGEN, &term_r_sem[asid], 0, 0);
+    SYSCALL(VERHOGEN, &swapSemaphore, 0, 0);
+    SYSCALL(VERHOGEN, &mainSemaphore, 0, 0);
     // Ammazzo il processo
     SYSCALL(TERMINATE, 0, 0, 0);
 }
